HTGCI-697: Replace Htg.Persistence ref with updated Htg.Common

+++ Beis.VendorManagement.Web/Models/Pricing/AdditionalDiscountsViewModel.cs
@@ -23,7 +23,7 @@
 
         public long ContractDurationDiscount { get; set; }
 
-        public long ContractDurationDiscountPercentage { get; set; }
+        public decimal ContractDurationDiscountPercentage { get; set; }
 
         public string ContractDurationDiscountDescription { get; set; }
